Fix: add submit_block_payload_info.hpp to protocol include path
